AI Developer Assignment/my project/requirements.txt
fastapi
uvicorn
sentence-transformers
PyMuPDF
faiss-cpu
pydantic
streamlit
python-multipart
